Generate TypeScript bindings for the system-api boundary with ts-rs
- new crate system_api_types holds the plain serde structs that cross
  ~system/BevyExplorerApi (moved from system_bridge, plus the
  PointerTargetType/Permission* enums from common and ClearableColor3
  from dcl, all re-exported from their old paths); ts-rs derives are
  unconditional and dcl_component's build.rs injects TS derives on the
  proto types the boundary embeds
- scripts/gen-ts-bindings.sh (called by `just ts-bindings` and CI)
  exports the bindings to react-web/src/engine/generated (gitignored)
- SetAvatarData now carries ClearableColor3 end-to-end instead of
  mapping to Option<Option<Color3>>: fixes "clear name color" decoding
  as "leave unchanged" across the native scene-process IPC, where
  Some(None) and None both encode as msgpack nil
- ProximityEvent.entity is now the u32 entity index rather than
  Entity::to_bits(), so the id survives JSON's f64 numbers; it is only
  used to pair enter/leave events
- 64-bit timestamp fields are overridden to `number` in the generated
  TS to match JSON.parse semantics
