+++ .gitignore
@@ -23,3 +23,6 @@ config.json
 /react-web/bridge-scene/static-native
 assets/react-hud/
 /assets/bridge-scene
+# ts-rs default export dir, written when the export_bindings tests run without
+# TS_RS_EXPORT_DIR (plain `cargo test`); the real output is scripts/gen-ts-bindings.sh's
+/crates/system_api_types/bindings
