+++ crates/common/Cargo.toml
@@ -11,6 +11,7 @@ dcl-assert = []
 [dependencies]
 platform = { workspace = true }
 dcl_component = { workspace = true }
+system_api_types = { workspace = true }
 
 bevy = { workspace = true }
 anyhow = { workspace = true }
