+++ crates/common/src/structs.rs
@@ -16,7 +16,7 @@ use bevy::{
 use dcl_component::proto_components::sdk::components::common::CameraTransition;
 use ethers_core::abi::Address;
 use serde::{Deserialize, Serialize};
-use strum_macros::EnumIter;
+pub use system_api_types::{PermissionLevel, PermissionType, PermissionValue, PointerTargetType};
 use tokio::sync::{OwnedSemaphorePermit, Semaphore};
 
 use crate::inputs::InputMapSerialized;
@@ -382,29 +382,6 @@ pub struct HeadSync {
     pub pitch_deg: f32,
     pub yaw_enabled: bool,
     pub pitch_enabled: bool,
-}
-
-/// Classifier for pointer-target hits — distinguishes scene world geometry,
-/// scene UI overlays, and avatars. Lives here (not in system_bridge) because
-/// the engine's pointer pipeline produces it before any scene API surface is
-/// involved; system_bridge re-exports the same value into its HoverEvent.
-#[derive(
-    Hash,
-    Clone,
-    Copy,
-    serde_repr::Serialize_repr,
-    serde_repr::Deserialize_repr,
-    Debug,
-    PartialEq,
-    Eq,
-    Default,
-)]
-#[repr(u32)]
-pub enum PointerTargetType {
-    #[default]
-    World = 0,
-    Ui = 1,
-    Avatar = 2,
 }
 
 /// World-space point-at state. On the local player, populated when the PointAt
@@ -877,32 +854,6 @@ pub struct SceneMeta {
     pub authoritative_multiplayer: Option<bool>,
 }
 
-#[derive(Serialize, Deserialize, Clone, Copy, PartialEq, Eq, Hash, Debug)]
-pub enum PermissionValue {
-    Allow,
-    Deny,
-    Ask,
-}
-
-#[derive(Serialize, Deserialize, Clone, Copy, PartialEq, Eq, Hash, Debug, EnumIter)]
-pub enum PermissionType {
-    MovePlayer,
-    ForceCamera,
-    PlayEmote,
-    SetLocomotion,
-    HideAvatarsNametags,
-    DisableVoice,
-    Teleport,
-    ChangeRealm,
-    SpawnPortable,
-    KillPortables,
-    Web3,
-    CopyToClipboard,
-    Fetch,
-    Websocket,
-    OpenUrl,
-}
-
 pub trait PermissionStrings {
     fn active(&self) -> &str;
     fn passive(&self) -> &str;
@@ -1015,13 +966,6 @@ impl PermissionStrings for PermissionType {
             PermissionType::CopyToClipboard => "copying text into the clipboard",
         }
     }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub enum PermissionLevel {
-    Scene(String),
-    Realm(String),
-    Global,
 }
 
 #[derive(Clone, Serialize, Deserialize, Event)]
