+++ crates/dcl_component/Cargo.toml
@@ -16,6 +16,7 @@ serde_json = { workspace = true }
 anyhow = { workspace = true }
 dcl-rpc = { workspace = true, optional = true }
 async-trait = { workspace = true }
+ts-rs = { workspace = true }
 
 [build-dependencies]
 prost-build = "0.11.8"
