+++ crates/system_api_types/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "system_api_types"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+
+[lints]
+workspace = true
+
+[dependencies]
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+serde_repr = "0.1"
+strum = { workspace = true }
+strum_macros = { workspace = true }
+dcl_component = { workspace = true }
+
+ts-rs = { workspace = true, features = ["serde-json-impl"] }
