+++ crates/system_bridge/Cargo.toml
@@ -22,6 +22,7 @@ ipfs = { workspace = true }
 serde_json = { workspace = true }
 dcl_component = { workspace = true }
 platform = { workspace = true }
+system_api_types = { workspace = true }
 
 serde_repr = "0.1"
 urlencoding = { workspace = true }
