+++ react-web/.gitignore
@@ -5,6 +5,9 @@ dist-ssr
 *.local
 .DS_Store
 
+# ts-rs bindings generated from crates/system_api_types (just ts-bindings)
+src/engine/generated
+
 # Playwright (tier-2 e2e) artifacts
 test-results
 playwright-report
